Move config to wxui
